Move LanguageRedirectContainer to common project
