+++ admin/resources/js/dashboard.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { LanguageRedirectContainer } from "./components/IntlContainer";
+import LanguageRedirectContainer from "@common/components/LanguageRedirectContainer";
 import { PoolDashboard } from "./components/PoolDashboard";
+import { loadLocaleData } from "./components/IntlContainer";
 
 ReactDOM.render(
-  <LanguageRedirectContainer>
+  <LanguageRedirectContainer getMessages={loadLocaleData}>
     <PoolDashboard />
   </LanguageRedirectContainer>,
   document.getElementById("app"),
