+++ common/src/components/LanguageRedirectContainer/index.ts
@@ -1,0 +1,4 @@
+export {
+  default,
+  LanguageRedirectContainer,
+} from "./LanguageRedirectContainer";
